pyproject.toml
[project]
name = "yan-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "yandex-cloud-ml-sdk>=0.12.0",
]
